Add grpc-web HTTP client
This client is only enabled when gRPC-Go is built with
GOARCH=WASM and GOOS=js. It is a spec-compliant
gRPC-Web client.
