+++ dial_js.go
@@ -1,0 +1,39 @@
+// +build js,wasm
+
+/*
+ *
+ * Copyright 2018 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package grpc
+
+import (
+	"errors"
+)
+
+var (
+	// errTransportSecurity indicates the client was attempting an RPC call
+	// from a WASM client with TransportCredentials configured.
+	errTransportSecurity = errors.New("cannot configure transport security for WASM clients")
+)
+
+func (cc *ClientConn) validateClientOptions() error {
+	if cc.dopts.copts.TransportCredentials != nil {
+		return errTransportSecurity
+	}
+
+	return nil
+}
